Additional comments by Harald Wagener
I did a comparison of the BLE devices back in February and this collection of reviews seems to be a better place to keep them than in a random G+ post (original source https://plus.google.com/+HaraldWagener/posts/YXAWNcduPjx). So I added the things that seem worthwhile to the reviews.